# reports/discovered-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ESignPay API",
    "description": "API for managing document e-signatures, payments (Stripe 3D Secure), account opening with KYC/OCR verification, and OTP signing via Twilio SMS.\n\n### Compression\nAll endpoints support **gzip compression**. Send `Accept-Encoding: gzip` to receive compressed responses (typically 60\u201380% payload reduction). Responses \u2265 256 bytes are automatically compressed.\n\n### Range / Partial Content (206)\nFile download endpoints support **HTTP Range requests** (`Range: bytes=start-end`). The server responds with `206 Partial Content` and `Accept-Ranges: bytes`, enabling resumable downloads and partial fetches for large files.",
    "contact": {
      "name": "ESignPay Team"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "ESignPay API Server \u2014 gzip compression enabled on all endpoints",
      "x-server-compression": {
        "algorithms": [
          "gzip"
        ],
        "mime-types": [
          "application/json",
          "application/xml",
          "text/html",
          "text/plain",
          "application/hal+json",
          "application/cbor"
        ],
        "enabled": true,
        "min-response-size": "256 bytes"
      },
      "x-server-range-support": {
        "enabled": true,
        "unit": "bytes",
        "endpoints": [
          "/api/documents/{id}/download"
        ]
      },
      "x-server-etag-support": {
...
      }
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/account-applications/{id}": {
      "get": {
        "tags": [
          "Account Applications"
        ],
        "summary": "Get account application detail (supports field filtering)",
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Comma-separated list of fields to include",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "description": "Previous ETag value. If it matches, server returns 304.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "\"0a1b2c3d4e5f\""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              },
              "ETag": {
                "description": "Entity tag for cache validation (MD5 of response body)",
                "schema": {
                  "type": "string",
                  "example": "\"0a1b2c3d4e5f\""
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAccountApplicationResponse"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified \u2014 returned when If-None-Match matches current ETag. No body is sent, saving bandwidth."
          }
        }
      },
      "put": {
        "tags": [
          "Account Applications"
        ],
        "summary": "Update an account application",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAccountApplicationResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Account Applications"
        ],
        "summary": "Delete an account application",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/account-applications": {
      "get": {
        "tags": [
          "Account Applications"
        ],
        "summary": "List my account applications (supports pagination & field filtering)",
        "operationId": "list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (0-based)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            },
            "example": 20
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Comma-separated list of fields to include (e.g. id,status,referenceNumber)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "description": "Previous ETag value. If it matches, server returns 304.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "\"0a1b2c3d4e5f\""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              },
              "ETag": {
                "description": "Entity tag for cache validation (MD5 of response body)",
                "schema": {
                  "type": "string",
                  "example": "\"0a1b2c3d4e5f\""
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageAccountApplicationResponse"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified \u2014 returned when If-None-Match matches current ETag. No body is sent, saving bandwidth."
          }
        }
      },
      "post": {
        "tags": [
          "Account Applications"
        ],
        "summary": "Create a new account application",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAccountApplicationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/account-applications/{id}/submit": {
      "post": {
        "tags": [
          "Account Applications"
        ],
        "summary": "Submit an account application for review",
        "operationId": "submit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAccountApplicationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/account-applications/{id}/regenerate-contract": {
      "post": {
        "tags": [
          "Account Applications"
        ],
        "summary": "Regenerate contract PDF",
        "operationId": "regenerateContract",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAccountApplicationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/account-applications/{id}/kyc": {
      "post": {
        "tags": [
          "Account Applications"
        ],
        "summary": "Upload a KYC document (base64)",
        "operationId": "uploadKyc",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadKycDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseKycDocumentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/account-applications/{id}/generate-contract": {
      "post": {
        "tags": [
          "Account Applications"
        ],
        "summary": "Generate contract PDF for the application",
        "operationId": "generateContract",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAccountApplicationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/account-types": {
      "get": {
        "tags": [
          "Account Applications"
        ],
        "summary": "List available account types (supports field filtering)",
        "operationId": "getAccountTypes",
        "parameters": [
          {
            "name": "fields",
            "in": "query",
            "description": "Comma-separated list of fields to include",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "description": "Previous ETag value. If it matches, server returns 304.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "\"0a1b2c3d4e5f\""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              },
              "ETag": {
                "description": "Entity tag for cache validation (MD5 of response body)",
                "schema": {
                  "type": "string",
                  "example": "\"0a1b2c3d4e5f\""
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListAccountTypeResponse"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified \u2014 returned when If-None-Match matches current ETag. No body is sent, saving bandwidth."
          }
        }
      }
    },
    "/api/account-applications/changes": {
      "get": {
        "tags": [
          "Account Applications"
        ],
        "summary": "Get applications changed since a given timestamp (delta sync)",
        "operationId": "getChanges",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "ISO date-time (e.g. 2026-04-01T00:00:00)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Comma-separated list of fields to include",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "description": "Previous ETag value. If it matches, server returns 304.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "\"0a1b2c3d4e5f\""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              },
              "ETag": {
                "description": "Entity tag for cache validation (MD5 of response body)",
                "schema": {
                  "type": "string",
                  "example": "\"0a1b2c3d4e5f\""
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListAccountApplicationResponse"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified \u2014 returned when If-None-Match matches current ETag. No body is sent, saving bandwidth."
          }
        }
      }
    },
    "/api/account-applications/{id}/kyc/{kycId}": {
      "delete": {
        "tags": [
          "Account Applications"
        ],
        "summary": "Delete a KYC document",
        "operationId": "deleteKyc",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "kycId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UpdateAccountApplicationRequest": {
        "type": "object",
        "properties": {
          "accountTypeCode": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "nationality": {
            "type": "string"
          },
          "addressLine1": {
            "type": "string"
          },
          "addressLine2": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "employmentStatus": {
            "type": "string"
          },
          "employerName": {
            "type": "string"
          },
          "jobTitle": {
            "type": "string"
          },
          "monthlyIncome": {
            "type": "number"
          }
        }
      },
...
      "AccountTypeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "monthlyFee": {
            "type": "number"
          }
        }
      },
      "ApiResponseAccountApplicationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AccountApplicationResponse"
          }
        }
      },
...
      "CreateAccountApplicationRequest": {
        "type": "object",
        "properties": {
          "accountTypeCode": {
            "type": "string",
            "minLength": 1
          },
          "dateOfBirth": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "nationality": {
            "type": "string"
          },
          "addressLine1": {
            "type": "string"
          },
          "addressLine2": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "employmentStatus": {
            "type": "string"
          },
          "employerName": {
            "type": "string"
          },
          "jobTitle": {
            "type": "string"
          },
          "monthlyIncome": {
            "type": "number"
          }
        },
        "required": [
          "accountTypeCode"
        ]
      },
      "UploadKycDocumentRequest": {
        "type": "object",
        "properties": {
          "documentType": {
            "type": "string",
            "minLength": 1
          },
          "fileName": {
            "type": "string",
            "minLength": 1
          },
          "contentType": {
            "type": "string",
            "minLength": 1
          },
          "fileContent": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "contentType",
          "documentType",
          "fileContent",
          "fileName"
        ]
      },
      "ApiResponseKycDocumentResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/KycDocumentResponse"
          }
        }
      },
...
      "ApiResponseListAccountTypeResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountTypeResponse"
            }
          }
        }
      },
      "ApiResponsePageAccountApplicationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageAccountApplicationResponse"
          }
        }
      },
...
      "ApiResponseListAccountApplicationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountApplicationResponse"
            }
          }
        }
      },
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {}
        }
      }
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Keycloak JWT Bearer token \u2014 obtain from POST to http://localhost:9090/realms/esignpayment/protocol/openid-connect/token",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}